refactor: 재사용없는 함수 제거, 코드 정리, 모달 파일위치 변경
